src/apis/flight-api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via flight-api (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
azure-core==1.35.0
    # via
    #   azure-cosmos
    #   azure-data-tables
    #   azure-identity
azure-cosmos==4.9.0
    # via flight-api (pyproject.toml)
azure-data-tables==12.7.0
    # via flight-api (pyproject.toml)
azure-identity==1.23.1
    # via flight-api (pyproject.toml)
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==45.0.5
    # via
    #   azure-identity
    #   msal
    #   pyjwt
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
exceptiongroup==1.3.0
    # via anyio
fastapi==0.116.1
    # via flight-api (pyproject.toml)
fastapi-cli==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
isodate==0.7.2
    # via azure-data-tables
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msal==1.33.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   fastapi-cloud-cli
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via msal
python-dotenv==1.1.1
    # via
    #   flight-api (pyproject.toml)
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.4
    # via
    #   azure-core
    #   msal
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.9
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.34.1
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via azure-core
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typer==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   azure-core
    #   azure-cosmos
    #   azure-data-tables
    #   azure-identity
    #   exceptiongroup
    #   fastapi
    #   multidict
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   flight-api (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via
    #   aiohttp
    #   azure-data-tables
